fix output of getSDpack

+++ platform/exacto/exactolink/exactolink_base.h
@@ -8,7 +8,7 @@
 #define EXACTOLINK_MLINE_SNSPACK_ID 0x17
 
 extern void exlnk_initSDpack(const uint8_t type, const uint16_t datalen, const uint32_t refcnt, ExactoBufferExtended * buffer);
-extern uint8_t exlnk_getSDpack(uint8_t * type, uint16_t * datalen, uint32_t * refcnt, ExactoBufferExtended * buffer);
+extern uint16_t exlnk_getSDpack(uint8_t * type, uint16_t * datalen, uint32_t * refcnt, ExactoBufferExtended * buffer);
 extern uint8_t exlnk_readsnsSDpack(const uint8_t type, uint8_t * sns_type, int16_t * dst, ExactoBufferExtended * buffer );
 extern uint8_t exlnk_pushtoSDpack(const uint8_t value, ExactoBufferExtended * buffer);
 extern uint8_t exlnk_pushSnsPack( const uint8_t sns_id, uint8_t * data, const uint16_t datacount,  ExactoBufferUint8Type * buffer);
